+++ proveedorPy/.env
@@ -1,6 +1,8 @@
 SERVIDOR-GRPC=localhost:9090
 DB_HOST=mysql-1d36c064-pato-ef11.a.aivencloud.com
 DB_PORT=25628
-DB_NAME=stk_proveedores
+DB_NAME=proveedores_stockearte
 DB_USER=avnadmin
 DB_PASSWORD=AVNS_ylqtAU8JPG7TNXz0mDD
+
+SERVER-KAFKA-BROKER=localhost:9092
