+++ proveedorPy/app/modulos/orden_de_compra/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+orden_de_compra_blueprint = Blueprint("orden_de_compra", __name__,static_folder='archivos')
+
+from . import OrdenDeCompraController
