+++ proveedorPy/app/static/archivos/ca.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEQTCCAqmgAwIBAgIULSXVETEUIDhS1eeDA3TFeDtEOB0wDQYJKoZIhvcNAQEM
+BQAwOjE4MDYGA1UEAwwvZTNmMTVhYjItYTc1Yy00MTI0LWIxYjEtMzA2MGZiNjUw
+ZDJkIFByb2plY3QgQ0EwHhcNMjQwMzI2MTE1MjU0WhcNMzQwMzI0MTE1MjU0WjA6
+MTgwNgYDVQQDDC9lM2YxNWFiMi1hNzVjLTQxMjQtYjFiMS0zMDYwZmI2NTBkMmQg
+UHJvamVjdCBDQTCCAaIwDQYJKoZIhvcNAQEBBQADggGPADCCAYoCggGBAKGEcmQV
+MQWggAJdH4MSEMWdmaLyYxBE6EPAOYU3fWZTyMJfWxMCwwgpNRBSyUoCHK7fc+a/
+MlxFWQ8AgpaRlIX0r+1pAlqrC2JRvWxCi1lOzYYTg9+N/13kvpT/yVMIVS8yZQVC
+led2lfUUUdqyGz9FDE6JzC2lXLYaYAoYeZGp1k/bi8qLUyGOVfeALWA3m1pCldDp
+zAIMHZDoSae1Jwb1DeoRW/7MAdEgmCKGGdBcv7J15ltDssFd8YySLtGZnuZHfhUI
+qqcvv1WpEqZ6VSOx4cYFGCTxEk0DfDlVHXKTl8dFN2OcBi6M+a6djYwD35oKDISJ
+jBq1ttj9XkwE5GPakpiGhlEafCoTuT+yjufXx+5GbT7ziflp416t6JCxMfg6Zn7E
+yg58NXDnOEvwVf2ae5d5lgiJUfaLJzk2jVWsp3W4iJixSGGQyD3F3cuAOGDiijIl
+beg14xoKVcMDyXzQecz7bIfTiMQAGQFvYltkDjT2bGfXLEhszpwoLAd19QIDAQAB
+oz8wPTAdBgNVHQ4EFgQU+DXrxvfrkx6aeIZ2UFvvDVeYhUYwDwYDVR0TBAgwBgEB
+/wIBADALBgNVHQ8EBAMCAQYwDQYJKoZIhvcNAQEMBQADggGBAJ9/dNcdpihpHLwG
+q9xIIpE3VH/pEgQzR2kkmEJ92eg1BYFAT8fEgoS7/GJZG8do15b2lfSXqLaNMilh
+CJacE0fYP8uc3oviZusEdKifBFGjWF+fBRSsbq90OEALgaiVIp7jZa7wlDOVsNbH
+2EKGCFUehhaId9lkVpZPSDtQYMV4f3QLavksBpYQE67a6aeewTARVf6H7kphPepw
+TvlZxAedhvwxy65J2qYow1HERSW/K8ki7AagWQmPD/6RZzYEtGUk0Ozln9ty2nnm
+51cgUbQv7WNdv8p30cDXGb9xyI5+DPwCbSN6YCE/JRATOD6gnlZibYjh+tdOrr/c
+/fhrePlkpFKkfxNwryaZixc49WoOg2RMGZq3Q5noWIwN4ilH/Yim0mjszthEeK7b
+5mVhiJmOz9p21QIdTtNtpz8ywMTqUMEtJStf3ROYIjzDaY3Vhw4DAM6wvd6EhzbU
+bG9lJBLLUwfSP+kzk4gWo3KPXQIaWshcOG1/kfXbPOAZAXYuAg==
+-----END CERTIFICATE-----
