+++ proveedorPy/app/static/archivos/service.cert
@@ -1,0 +1,26 @@
+-----BEGIN CERTIFICATE-----
+MIIEYDCCAsigAwIBAgIUFK9/XqGRhSqcVi8iwLItpE5qqU4wDQYJKoZIhvcNAQEM
+BQAwOjE4MDYGA1UEAwwvZTNmMTVhYjItYTc1Yy00MTI0LWIxYjEtMzA2MGZiNjUw
+ZDJkIFByb2plY3QgQ0EwHhcNMjQwOTMwMDEzMDQ2WhcNMjYxMjI5MDEzMDQ2WjA+
+MRYwFAYDVQQKDA1rYWZrYS04NzFhNTc4MREwDwYDVQQLDAh1NHkxbTZ6NjERMA8G
+A1UEAwwIYXZuYWRtaW4wggGiMA0GCSqGSIb3DQEBAQUAA4IBjwAwggGKAoIBgQC6
+JlAoqF+BB5+rEXDzgUvgz1WgkBnb+uydU4LkkRk99Fi4sQpG3y5Tof8DBEOycJN8
+qa0H9zbBlRRSaJTOhWd+ou8O1wXsAYiWFrGurQBPO8lS7FzbwDhyuBQOdLnBJiP+
+m2RYCoKPHpe70NZ+WiWb9NAZ3UuSGl281HU7cVYlIpZZIQiW6lo8Yjd1T8oIGMRp
+ntp8F7BS4ahPArtMCMt1z82Jst3GMIAtjVtRiNsmtLlpnBiOsT0+jfX2Ez9IfrAO
+8t9VAFkxtc+7B82dsp65Xft8uHQUMEgM8CiJ2eJ4HOApMd4DNK/bTca+rIjBZrt+
+ZLKc7SwpbPpHsbVIXupIFILNoQGctY5hFAaOmHM8DXRvx0SXA7KbQIQQMNkB5O+p
+X3MUmF5eZ1v2zW8JzVRnexxcvIt3fjvoHPsvGLNZR0cjzH66Wfth2/+Ga42J88xd
+xwa+TCccO81AO0u6v2ItsY3zfrK36iXiSuBguEcR4Y8t/HScG/5DZRw94OzYX4UC
+AwEAAaNaMFgwHQYDVR0OBBYEFIpiiNBBxS4ICFowdxmgwjPbLWo0MAkGA1UdEwQC
+MAAwCwYDVR0PBAQDAgWgMB8GA1UdIwQYMBaAFPg168b365MemniGdlBb7w1XmIVG
+MA0GCSqGSIb3DQEBDAUAA4IBgQCQ8peqQ+NaCZC32U7/IwxSr1WOUFbHT6dO3YfW
+v4QvK84PxPVMYS65qcYHgAL1e24gtMtI6YiBclsKCbuVWsMS0kqaf3PcGOp5NzIA
+MfSMXdrOZH00j9Aa3Kh439IYB9M24EroRT/H8Fw8ixnG5eGty/gsBrtxouXQMXOm
+wwEInAvneo6Tiimv/VFZMi9yGRhxUcCJUhEeoen2hfxo2vMrblYzauPSi8hR+DRs
+OoWaJC6nz4ofhUQ1bw+wNN7bGtsMOCAK214JXKtx8ILI/Mdhh1YZquK0KNIZDOYP
+1k8Qg7MdDKF6fqVvxTcwod1V2Pk+H5FLcncRJRx3rPb0xcaYGhov5DGTzImo9D8A
+h8wX6Toh/F7tSrz6Lj9x9e9+dC5oLqOsj4cYzH+ntSLYDMAQEO2KhMEw7cah70JI
+417MxKPybfmHh9IdjNLu19MCY2BMCcKWt0yeB46TqxeSNNjUzvADno9q3TIqa1h3
+8nBeMev4s+roVLVnlc4jVuEGsBQ=
+-----END CERTIFICATE-----
