+++ proveedorPy/app/static/archivos/service.key
@@ -1,0 +1,40 @@
+-----BEGIN PRIVATE KEY-----
+MIIG/QIBADANBgkqhkiG9w0BAQEFAASCBucwggbjAgEAAoIBgQC6JlAoqF+BB5+r
+EXDzgUvgz1WgkBnb+uydU4LkkRk99Fi4sQpG3y5Tof8DBEOycJN8qa0H9zbBlRRS
+aJTOhWd+ou8O1wXsAYiWFrGurQBPO8lS7FzbwDhyuBQOdLnBJiP+m2RYCoKPHpe7
+0NZ+WiWb9NAZ3UuSGl281HU7cVYlIpZZIQiW6lo8Yjd1T8oIGMRpntp8F7BS4ahP
+ArtMCMt1z82Jst3GMIAtjVtRiNsmtLlpnBiOsT0+jfX2Ez9IfrAO8t9VAFkxtc+7
+B82dsp65Xft8uHQUMEgM8CiJ2eJ4HOApMd4DNK/bTca+rIjBZrt+ZLKc7SwpbPpH
+sbVIXupIFILNoQGctY5hFAaOmHM8DXRvx0SXA7KbQIQQMNkB5O+pX3MUmF5eZ1v2
+zW8JzVRnexxcvIt3fjvoHPsvGLNZR0cjzH66Wfth2/+Ga42J88xdxwa+TCccO81A
+O0u6v2ItsY3zfrK36iXiSuBguEcR4Y8t/HScG/5DZRw94OzYX4UCAwEAAQKCAYA9
+izg6UPyOWNppwRDyeP+AupvJQ4tBJxM+qqMedXh+esAwg9vtdZFH9kmtEiZdowmo
+ECMUfJ/eXa8D1e7I3uNKQ408/0tGbXVehe9F3I27JH2ZzGqg/srK5uA+rUnSuYzx
++ycwUn062zb4UgHwcVZ42vZIbtky20O9kBKYTg7YpThHDHxeBOpDh4dlYjVd+u5J
+K0t59o0FwfSW9sS8zTWTZFpEmd3nzhY6uT8FVHV8NQkOCR03ZBLnESFOM6vbCIMU
+XyiN08SURml5EPS1PRqPCaO+oIjDBNNdf5BhKUf/jHmhWsH0qHpPgwlgJkNPtqTh
+f7C7ILmm/r4Wb9KlHo0euCayHFswzXrthH7Nf/UhULI+jDDdJ35C1ZZ3IGsYM831
+J1JMCXyZqG8ALfSfaUGcQvgvMCvYPgwrOoZMUS9O+L0FyGLiw5XSyYCiKzsUAJTr
+MpTsFydX5KDhPFKFm6svMThOZ6A4UtmpEpvAjpswy25Q9alxjLzK00izjB9eQnsC
+gcEA9zMJbrglIX6e9nRLMVk22jA1MegtkOTGL1keCIJuZyW/KLFT0gF83EBytPUG
+yKGgYQ+Xd4Z60iPrQXWRfrAaFsRUTeiAq6WkGgpb8l8itsQ1AZHnY/4A0mt/OWAY
+/+EsaBgRpun1ozJiMJucGjBHJbAqE7sLmxyu9OsRVfOY5b3g7UxWxHleNZBNbMrR
+9+cQaAnswighAxZkUXRXVIBuY3JCrFEqfp/e1yEjSgyCe7iNDpaQBMGcu3wU/0wp
+psNTAoHBAMDG30ljRcF1FgNinbUgIqD0AIxrKJ7XJjmkFg69pQ2QHkrd2bgvtVCq
+XiS33NQvf402jMdBH7UgcB+Mqdo+vWpexioAtoU+NY/v0UfTpsQz4rvMQOqLlQ78
+K4fN8P27pHpDzR4HD5F1Xv9Nd6M/xczS8h4vMXzPtUhrcKcMwdZlyIHHz8iq/CYb
+UgZlDvgnCQH3lYCIMo6oYxF2oj076ZPhyOkeuHuDEnNKbr9p2Xh6yji6v7hhKiRK
+sjPcc1gOxwKBwQDN00j/T4RbJzrkqmARbt8lPQ6Nx2xpQJNjySggnbmfWOaZ00yf
+h22XVnIsCfjkBO+s4XaUpvyNDCtODb5fZfHTExW+6OB6CylgJU/5tw0IcMZR0FKC
+B3RXv/nmpAUsyShYY/IOsK8+ey0Wm/qWM6yJzOy8EPnQ5Ukn1LJShsnAQSGPdAiq
+/wNaRhSxiQ7bAyG9/sWLxC34jCRrThuI2IcRdondQJfJZwP8zDK97TyYF0ob43F3
+BF6kHOS+zYPAvtECgcB2/st5u4OlhSbPdCIxWeLGQ5fLFaJ/WgNIdVcARmQJ2To5
+sHPiU+CyKFfwhmYTvDYZ7uA7ZqGx5pIPh36ZLz0w2cusxb/kw5/Ad6JRmf5JAlYk
+4lH39a1PEK22KJNnrFo89nk43dboTPEU2rbM2GK6qkO2Ch+cEYKvb1Kcay2DsFRy
+d+p0zcbTpu2c8OiqmT3gLvnj2y6KDLsKXpaLAd6O4EJqVAJAi4nqDLoMAcSxr+n/
+1pbds2yV9l3F0YCJjKUCgcAlv0SBQgEYZvBh51QPw2InUOi90sbBi61cL1eSStsr
+7CJBMrJKSxfQauCSWnO7xSP0QvVbX2nlYDiyQ3+CHuQUcg/AFwsmWrtmTYZrjAOD
+juy1Yzt5ODi9NTPAeFaxJSh59BTyDXud7Bga1TQaYuEpChL7wDnYlEVxHi9MY1ZR
+zZotBMIh2x8aK3opUSaTsTm0FRFaQOKvb/Lgi4CjcfQ1ZX9uLdox1gakpIelWy/3
+SWZURo5rUUtcpa5hyYwe08A=
+-----END PRIVATE KEY-----
