+++ proveedorPy/app/templates/home.html
@@ -26,6 +26,16 @@
             </div>
         </div>
     </div>
+    <div class="col">
+        <div class="card" style="width: 18rem;">
+            <img src="https://manada.com.ar/blog/wp-content/uploads/2016/04/fashion-winter.png" class="card-img-top" alt="..." width="100%" height="200">
+            <div class="card-body">
+                <h5 class="card-title">CONSUMER ORDEN DE COMPRA</h5>
+                <p class="card-text">Guarda en la DB todas las ODC producidas</p>
+                <a href="/orden_de_compra" class="btn btn-primary">START CONSUMER ODC</a>
+            </div>
+        </div>
+    </div>
 </div>
 
 </div>
