+++ proveedorPy/proveedor_py.sql
@@ -31,9 +31,10 @@ CREATE TABLE `orden_de_compra` (
   `id` INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,
   `observaciones` TEXT DEFAULT NULL,
   `id_store` int DEFAULT NULL,
+  `store_code` varchar(255) DEFAULT NULL,
   `estado` varchar(255) DEFAULT 'SOLICITADA',
-  `fecha_creacion` DATETIME DEFAULT NULL,
-  `fecha` DATETIME DEFAULT NULL,
+  `fecha_solicitud` DATETIME DEFAULT NULL,
+  `fecha_procesamiento` DATETIME DEFAULT NULL,
   `fecha_recepcion` DATETIME DEFAULT NULL,
   `id_despacho` int DEFAULT NULL,
    `procesado` int DEFAULT 0,
